requirements.txt
selenium==4.9.0
undetected-chromedriver
python-dotenv
